feedback: refactor file name and location for string constants.
