+++ test/pages/signing_official_console/SigningOfficialDarApprovals.spec.tsx
@@ -7,7 +7,7 @@ import { Collections } from 'src/libs/ajax/Collections'
 import { USER_ROLES } from 'src/libs/utils'
 import { consoleTypes } from 'src/utils/DarCollectionUtils'
 import { DarCollectionSummary } from 'src/types/model'
-import { PI_QUALIFICATION } from 'src/libs/principalInvestigator'
+import { PI_QUALIFICATION } from 'src/definitions/definitions-en-us'
 
 type MockDarCollectionTableProps = {
   collections: DarCollectionSummary[]
